+++ devtools_m1/cypress/integration/common/admin.js
@@ -1,5 +1,5 @@
 import { Given, When, Then } from "cypress-cucumber-preprocessor/steps"
-import { mapFrontendWebsiteId } from "../../lib/frontend_context"
+import { mapFrontendWebsiteId, mapFrontendStoreViewId } from "../../lib/frontend_context"
 
 Given('I am logged into the admin interface', function() {
   cy.visit(`http://main.magento.localhost:3005/index.php/admin`)
@@ -66,6 +66,7 @@ Given('I have configured Drip to be enabled for {string}', function(scope) {
 // Simple Product
 Given('I have configured a simple widget for {string}', function(site) {
   cy.createProduct({
+    "storeId": mapFrontendStoreViewId(site),
     "sku": "widg-1",
     "name": "Widget 1",
     "description": "This is really a widget. There are many like it, but this one is mine.",
@@ -82,6 +83,7 @@ Given('I have configured a configurable widget', function() {
     "description": "This is really a widget. There are many like it, but this one is mine.",
     "shortDescription": "This is really a widget.",
     "typeId": "configurable",
+    "image": "parent_image.png",
     "attributes": {
       "widget_size": {
         "XL": {
@@ -101,6 +103,7 @@ Given('I have configured a configurable widget', function() {
   })
 })
 
+
 // TODO: Consider merging this and the above.
 Given('I have configured a configurable widget for website {string}', function(site) {
   const websiteId = mapFrontendWebsiteId(site)
@@ -119,14 +122,43 @@ Given('I have configured a configurable widget for website {string}', function(s
           "name": "Widget 1 XL",
           "description": "This is really an XL widget. There are many like it, but this one is mine.",
           "shortDescription": "This is really an XL widget.",
-          "websiteIds": [websiteId],
+          "websiteIds": [websiteId]
         },
         "L": {
           "sku": "widg-1-l",
           "name": "Widget 1 L",
           "description": "This is really an L widget. There are many like it, but this one is mine.",
           "shortDescription": "This is really an L widget.",
-          "websiteIds": [websiteId],
+          "websiteIds": [websiteId]
+        }
+      }
+    }
+  })
+})
+
+Given('I have configured a configurable widget with an invisible child', function() {
+  cy.createProduct({
+    "sku": "widg-1",
+    "name": "Widget 1",
+    "description": "This is really a widget. There are many like it, but this one is mine.",
+    "shortDescription": "This is really a widget.",
+    "image": "parent_image.png",
+    "typeId": "configurable",
+    "attributes": {
+      "widget_size": {
+        "XL": {
+          "sku": "widg-1-xl",
+          "name": "Widget 1 XL",
+          "description": "This is really an XL widget. There are many like it, but this one is mine.",
+          "shortDescription": "This is really an XL widget.",
+          "visibility": 1
+        },
+        "L": {
+          "sku": "widg-1-l",
+          "name": "Widget 1 L",
+          "description": "This is really an L widget. There are many like it, but this one is mine.",
+          "shortDescription": "This is really an L widget.",
+          "visibility": 1
         }
       }
     }