+++ devtools_m1/php_utils/create_product.php
@@ -50,6 +50,7 @@ class Mage_Shell_Drip_CreateProduct extends Mage_Shell_Abstract
             "taxClassId" => 0, //tax class (0 - none, 1 - default, 2 - taxable, 4 - shipping)
             "price" => 11.22,
             "cost" => 22.33,
+            "image" => "my_image.png",
             "attributeSetId" => $defaultAttrSetId,
             "createdAt" => strtotime('now'),
             "updatedAt" => strtotime('now'),
@@ -59,6 +60,7 @@ class Mage_Shell_Drip_CreateProduct extends Mage_Shell_Abstract
                 "is_in_stock" => 1,
                 "qty" => 999
             ),
+            "visibility" => Mage_Catalog_Model_Product_Visibility::VISIBILITY_BOTH, //catalog and search visibility
         );
         $fullData = array_replace_recursive($defaults, $data);
 
@@ -67,8 +69,6 @@ class Mage_Shell_Drip_CreateProduct extends Mage_Shell_Abstract
             $methodName = "set".ucfirst($key);
             $product->$methodName($value);
         }
-
-        $product->setVisibility(Mage_Catalog_Model_Product_Visibility::VISIBILITY_BOTH); //catalog and search visibility
 
         return $product;
     }